isolate prototypes in their own module
