+++ drools-model/drools-canonical-model/src/main/java/org/drools/model/Variable.java
@@ -25,4 +25,8 @@ public interface Variable<T> extends Argument<T> {
     default boolean isFact() {
         return true;
     }
+
+    default boolean isPrototype() {
+        return false;
+    }
 }
