+++ drools-model/drools-model-codegen/pom.xml
@@ -77,6 +77,10 @@
             <groupId>org.drools</groupId>
             <artifactId>drools-model-compiler</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.drools</groupId>
+            <artifactId>drools-model-prototype</artifactId>
+        </dependency>
 
         <dependency>
             <groupId>com.github.javaparser</groupId>
