+++ drools-model/drools-model-codegen/src/test/java/org/drools/model/codegen/execmodel/FactTemplateTest.java
@@ -18,6 +18,17 @@
  */
 package org.drools.model.codegen.execmodel;
 
+import java.math.BigDecimal;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.concurrent.TimeUnit;
+import java.util.function.BiPredicate;
+import java.util.stream.Collectors;
+
 import org.drools.base.facttemplates.Event;
 import org.drools.base.facttemplates.Fact;
 import org.drools.base.facttemplates.FactTemplateObjectType;
@@ -30,10 +41,6 @@ import org.drools.model.ConstraintOperator;
 import org.drools.model.DSL;
 import org.drools.model.Index;
 import org.drools.model.Model;
-import org.drools.model.Prototype;
-import org.drools.model.PrototypeExpression;
-import org.drools.model.PrototypeFact;
-import org.drools.model.PrototypeVariable;
 import org.drools.model.Query;
 import org.drools.model.Rule;
 import org.drools.model.Variable;
@@ -41,6 +48,10 @@ import org.drools.model.codegen.execmodel.domain.Person;
 import org.drools.model.codegen.execmodel.domain.Result;
 import org.drools.model.functions.Function1;
 import org.drools.model.impl.ModelImpl;
+import org.drools.model.prototype.Prototype;
+import org.drools.model.prototype.PrototypeExpression;
+import org.drools.model.prototype.PrototypeFact;
+import org.drools.model.prototype.PrototypeVariable;
 import org.drools.modelcompiler.KieBaseBuilder;
 import org.junit.Test;
 import org.kie.api.KieBase;
@@ -53,17 +64,6 @@ import org.kie.api.runtime.rule.FactHandle;
 import org.kie.api.runtime.rule.Row;
 import org.kie.api.runtime.rule.ViewChangedEventListener;
 import org.kie.api.time.SessionPseudoClock;
-
-import java.math.BigDecimal;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.concurrent.TimeUnit;
-import java.util.function.BiPredicate;
-import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.fail;
@@ -80,16 +80,16 @@ import static org.drools.model.PatternDSL.pattern;
 import static org.drools.model.PatternDSL.query;
 import static org.drools.model.PatternDSL.reactOn;
 import static org.drools.model.PatternDSL.rule;
-import static org.drools.model.PrototypeDSL.field;
-import static org.drools.model.PrototypeDSL.protoPattern;
-import static org.drools.model.PrototypeDSL.prototype;
-import static org.drools.model.PrototypeDSL.variable;
-import static org.drools.model.PrototypeExpression.fixedValue;
-import static org.drools.model.PrototypeExpression.prototypeField;
-import static org.drools.model.PrototypeExpression.thisPrototype;
 import static org.drools.model.codegen.execmodel.BaseModelTest.getObjectsIntoList;
-import static org.drools.modelcompiler.facttemplate.FactFactory.createMapBasedEvent;
-import static org.drools.modelcompiler.facttemplate.FactFactory.createMapBasedFact;
+import static org.drools.model.prototype.PrototypeDSL.field;
+import static org.drools.model.prototype.PrototypeDSL.protoPattern;
+import static org.drools.model.prototype.PrototypeDSL.prototype;
+import static org.drools.model.prototype.PrototypeDSL.variable;
+import static org.drools.model.prototype.PrototypeExpression.fixedValue;
+import static org.drools.model.prototype.PrototypeExpression.prototypeField;
+import static org.drools.model.prototype.PrototypeExpression.thisPrototype;
+import static org.drools.model.prototype.facttemplate.FactFactory.createMapBasedEvent;
+import static org.drools.model.prototype.facttemplate.FactFactory.createMapBasedFact;
 
 public class FactTemplateTest {
