--- drools-model/drools-model-compiler/src/main/resources/META-INF/services/org.drools.model.PrototypeFactFactory
@@ -1,1 +1,0 @@
-org.drools.modelcompiler.facttemplate.PrototypeFactFactoryImpl
