+++ drools-model/drools-model-compiler/src/main/resources/META-INF/services/org.drools.model.prototype.PrototypeFactFactory
@@ -1,0 +1,1 @@
+org.drools.model.prototype.facttemplate.PrototypeFactFactoryImpl
