+++ drools-model/drools-model-prototype/src/main/java/org/drools/model/prototype/PrototypeExpression.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.drools.model;
+package org.drools.model.prototype;
 
 import java.util.Arrays;
 import java.util.Collection;
