+++ drools-model/drools-model-prototype/src/main/java/org/drools/model/prototype/facttemplate/HashMapEventImpl.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.drools.modelcompiler.facttemplate;
+package org.drools.model.prototype.facttemplate;
 
 import java.util.HashMap;
 import java.util.Map;
