+++ drools-model/drools-model-prototype/src/main/resources/META-INF/services/org.drools.modelcompiler.PrototypeService
@@ -1,0 +1,1 @@
+org.drools.model.prototype.impl.PrototypeServiceImpl
