+++ drools-model/pom.xml
@@ -39,9 +39,10 @@
     <module>drools-codegen-common</module>
     <module>drools-canonical-model</module>
     <module>drools-model-compiler</module>
-    <module>drools-model-codegen</module>
     <module>drools-mvel-parser</module>
     <module>drools-mvel-compiler</module>
+    <module>drools-model-prototype</module>
+    <module>drools-model-codegen</module>
   </modules>
