+++ drools-reliability/drools-reliability-tests/src/test/java/org/drools/reliability/test/ReliabilityCepTimeWindowTest.java
@@ -25,7 +25,7 @@ import java.util.concurrent.TimeUnit;
 import org.drools.model.Global;
 import org.drools.model.Index;
 import org.drools.model.Model;
-import org.drools.model.PrototypeVariable;
+import org.drools.model.prototype.PrototypeVariable;
 import org.drools.model.Rule;
 import org.drools.model.impl.ModelImpl;
 import org.drools.reliability.test.util.TimeAmount;
@@ -41,8 +41,8 @@ import static org.drools.model.DSL.after;
 import static org.drools.model.DSL.globalOf;
 import static org.drools.model.DSL.on;
 import static org.drools.model.PatternDSL.rule;
-import static org.drools.model.PrototypeDSL.protoPattern;
-import static org.drools.model.PrototypeDSL.variable;
+import static org.drools.model.prototype.PrototypeDSL.protoPattern;
+import static org.drools.model.prototype.PrototypeDSL.variable;
 import static org.drools.reliability.test.util.PrototypeUtils.DEFAULT_PROTOTYPE_NAME;
 import static org.drools.reliability.test.util.PrototypeUtils.getPrototype;
